Fix merge conflicts and move doGet to lessonhandler
