+++ capstone/src/main/java/com/google/sps/servlets/LessonHandlerServlet.java
@@ -8,6 +8,14 @@ import com.google.appengine.api.datastore.DatastoreServiceFactory;
 import com.google.appengine.api.datastore.PreparedQuery;
 import com.google.appengine.api.datastore.Query;
 import com.google.appengine.api.datastore.Entity;
+import com.google.appengine.api.datastore.Key;
+import com.google.appengine.api.datastore.EntityNotFoundException;
+import java.util.List;
+import java.util.ArrayList;
+import com.google.gson.Gson;
+import com.google.sps.data.Lesson;
+import com.google.sps.data.Room;
+import com.google.sps.data.RequestParser;
 import com.google.sps.service.DatabaseService;
 import com.google.sps.data.RequestParser;
 import javax.servlet.annotation.WebServlet;
@@ -17,6 +25,22 @@ import javax.servlet.http.HttpServletResponse;
 
 @WebServlet("/lesson")
 public class LessonHandlerServlet extends HttpServlet {
+
+    @Override
+    public void doGet(HttpServletRequest request, HttpServletResponse response) throws IOException {
+        DatastoreService datastore = DatastoreServiceFactory.getDatastoreService();
+        PreparedQuery roomResults = datastore.prepare(new Query(Room.ROOM_ENTITY_NAME));
+        Room room = null;
+
+        for(Entity entity : roomResults.asIterable()) {
+            if (new Room(entity).getRoomKey().getId() == Long.parseLong(request.getParameter("room_id"))) {
+                room = new Room(entity);
+            }
+        }
+        
+        response.setContentType("application/json");
+        response.getWriter().println(new Gson().toJson(getLessons(room)));
+    }
 
     @Override
     public void doPost(HttpServletRequest request, HttpServletResponse response) throws IOException {
@@ -36,5 +60,19 @@ public class LessonHandlerServlet extends HttpServlet {
         
         DatabaseService.save(room.getRoomEntity()); 
         response.setStatus(HttpServletResponse.SC_OK);
+    }
+
+    public List<Lesson> getLessons(Room room) {
+        List<Lesson> lessons = new ArrayList<>();
+    
+        try {
+            for(Key key : room.getAllLessons()) {
+                lessons.add(DatabaseService.getLesson(key));
+            }
+        }
+        catch (EntityNotFoundException e) {
+            System.out.println("Lesson entities don't exist.");
+        }
+        return lessons;
     }
 }
